task: Split copy into copy_file and copy_stream
