--- task/stream.go
@@ -1,1 +1,0 @@
-package task
